new changes-2 (views.py and urls.py in Pages app)
