+++ templates/Pages/contact.html
@@ -30,7 +30,7 @@
         {% endfor %}
         <button type="reset" class="btn btn-secondary">Clear</button>
         <button type="submit" class="btn btn-primary">Submit</button><br><br>
-        <a class="btn btn-outline-dark" href="{% url 'Product:product-list' %}">Back to Home</a>
+        <a class="btn btn-outline-dark" href="{% url 'Products:product-list' %}">Back to Home</a>
     </form>
 </div>
