+++ templates/Pages/register.html
@@ -10,7 +10,7 @@
                 <button class="btn btn-outline-info" style="color: red" type="submit">Register</button>
             </form>
 
-            <p style="margin-top: 5%; font-size: 75%;">If you have already account. <a href="{% url 'Product:login-request' %}"><strong>Login</strong></a></p>
+            <p style="margin-top: 5%; font-size: 75%;">If you have already account. <a href="{% url 'Auth:login-request' %}"><strong>Login</strong></a></p>
         </div>
     </div>
 </div>
