+++ templates/ProductPages/includes/categories.html
@@ -1,5 +1,5 @@
 <div class="row">
-    {% for category in object_list_c %}
+    {% for category in category_object_list %}
     <div class="col-lg-2">
         <a class="btn" href="{% url 'Products:product-list' category.id %}">
             <div class="box-element product ">
