+++ templates/ProductPages/includes/footer.html
@@ -13,7 +13,7 @@
                     <li>
                         <a class="grey-text text-lighten-3" href="#!"><img id="about-us-icon" src="{% static 'images/about-us.png' %}"> Biz haqimizda</a>
                     </li>
-                    <li><a class="grey-text text-lighten-3" href="{% url 'Products:contact-request' %}">&#9993; Bog'lanish</a></li>
+                    <li><a class="grey-text text-lighten-3" href="{% url 'Auth:contact-request' %}">&#9993; Bog'lanish</a></li>
                     <li>
                         <a class="grey-text text-lighten-3" href="#!"><img id="security-icon" src="{% static 'images/security.png' %}"> Maxfiylik Siyosati</a>
                     </li>
