+++ templates/ProductPages/includes/navbar.html
@@ -17,11 +17,11 @@
 
             {% if user.is_staff %}
             <a href="/accounts" class="btn btn-outline-success" style="color: black;">Welcome {{ user.username }}</a>
-            <a href="{% url 'Products:logout-request' %}" class="btn btn-outline-warning" style="color: black;">Logout</a>
+            <a href="{% url 'Auth:logout-request' %}" class="btn btn-outline-warning" style="color: black;">Logout</a>
             <a href="{% url 'Products:product-create' %}" class="btn btn-outline-info" style="color: black;">Create</a> {% elif user.is_authenticated %}
             <a href="" class="btn btn-outline-success" style="color: rgb(70, 3, 3);">Welcome {{ user.username }}</a>
-            <a href="{% url 'Products:logout-request' %}" class="btn btn-outline-warning" style="color: black;">Logout</a> {% else %}
-            <a href="{% url 'Products:login-request' %}" class="btn btn-warning" style="color: black;">Login</a> {% endif %}
+            <a href="{% url 'Auth:logout-request' %}" class="btn btn-outline-warning" style="color: black;">Logout</a> {% else %}
+            <a href="{% url 'Auth:login-request' %}" class="btn btn-warning" style="color: black;">Login</a> {% endif %}
 
             <a href="#">
                 <img id="cart-icon" src="{% static 'images/cart.png' %}">
